requirements.txt
Django~=5.1.2
djangorestframework~=3.14.0
faker~=18.3.0
django-cors-headers~=4.9.0
django-filter~=24.3

dj-database-url==2.2.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
whitenoise==6.7.0